test-hl: fix MSVC C4334 on the flow-control budget shift
1u << m computes in 32 bits and implicitly widens to the 64-bit
size_t budget, which MSVC flags as C4334 under -Werror.  Shift in
the target width instead, matching the existing (size_t)1 << m use
at :955 in the same file.

+++ test-apps/test-hl.c
@@ -504,7 +504,7 @@ run_goldens(const struct golden *gtab, size_t gcount,
 		 */
 
 		for (m = 0; m < 8; m++) {
-			size_t budget = 1u << m;
+			size_t budget = (size_t)1 << m;
 
 			ret = run_deferred(ops, (const uint8_t *)g->in,
 					   strlen(g->in), budget, cap);
